Trace IDs should be 32 bytes of random hex chars
The function as-is will always generate a trace ID that ends in 0. The reason is that the array passed to crypto.getRandomValues() only has 31 elements. That's fine when you're looking at the GUID function, which only has 31 random characters.

In the current code, the array is created with 31 elements, indexed 0-30. The last character is pulled using `randomVals[31]` which returns undefined, which gets translated to 0 (`undefined & 15 === 0`).